+++ rmk/src/config/mod.rs
@@ -1,6 +1,8 @@
 mod behavior;
 #[cfg(feature = "_ble")]
 mod ble_battery;
+#[cfg(feature = "_ble")]
+mod ble_host_power;
 mod device;
 mod positional;
 mod storage;
@@ -12,6 +14,8 @@ pub use behavior::{
 };
 #[cfg(feature = "_ble")]
 pub use ble_battery::BleBatteryConfig;
+#[cfg(feature = "_ble")]
+pub use ble_host_power::BleHostPowerConfig;
 pub use device::{DeviceConfig, RMK_BUILD_INFO};
 pub use positional::{Hand, PositionalConfig};
 pub use storage::StorageConfig;
@@ -27,4 +31,6 @@ pub struct RmkConfig<'a> {
     pub storage_config: StorageConfig,
     #[cfg(feature = "_ble")]
     pub ble_battery_config: BleBatteryConfig<'a>,
+    #[cfg(feature = "_ble")]
+    pub ble_host_power_config: Option<BleHostPowerConfig>,
 }